Add datastore basic roundtrip testing and fix some issues with the datastores.

+++ domain/consensus/datastructures/headersselectedchainstore/headers_selected_chain_staging_shard.go
@@ -1,6 +1,8 @@
 package headersselectedchainstore
 
 import (
+	"fmt"
+
 	"github.com/Hoosat-Oy/HTND/domain/consensus/database/binaryserialization"
 	"github.com/Hoosat-Oy/HTND/domain/consensus/model"
 	"github.com/Hoosat-Oy/HTND/domain/consensus/model/externalapi"
@@ -31,6 +33,21 @@ func (hscss *headersSelectedChainStagingShard) Commit(dbTx model.DBTransaction) 
 		return nil
 	}
 
+	prevHighest, prevExists, err := hscss.store.highestChainBlockIndex(dbTx)
+	if err != nil {
+		return err
+	}
+	prevLength := uint64(0)
+	if prevExists {
+		prevLength = prevHighest + 1
+	}
+	removedCount := uint64(len(hscss.removedByIndex))
+	addedCount := uint64(len(hscss.addedByIndex))
+	if prevLength < removedCount {
+		return fmt.Errorf("headersSelectedChainStagingShard: removed %d entries from chain of length %d", removedCount, prevLength)
+	}
+	newLength := prevLength - removedCount + addedCount
+
 	for hash := range hscss.removedByHash {
 		hashCopy := hash
 		err := dbTx.Delete(hscss.store.hashAsKey(&hashCopy))
@@ -48,7 +65,6 @@ func (hscss *headersSelectedChainStagingShard) Commit(dbTx model.DBTransaction) 
 		hscss.store.cacheByIndex.Remove(index)
 	}
 
-	highestIndex := uint64(0)
 	for hash, index := range hscss.addedByHash {
 		hashCopy := hash
 		err := dbTx.Put(hscss.store.hashAsKey(&hashCopy), hscss.store.serializeIndex(index))
@@ -64,17 +80,24 @@ func (hscss *headersSelectedChainStagingShard) Commit(dbTx model.DBTransaction) 
 		hscss.store.cacheByHash.Add(&hashCopy, index)
 		hscss.store.cacheByIndex.Add(index, &hashCopy)
 
-		if index > highestIndex {
-			highestIndex = index
-		}
 	}
 
-	err := dbTx.Put(hscss.store.highestChainBlockIndexKey, hscss.store.serializeIndex(highestIndex))
+	if newLength == 0 {
+		err := dbTx.Delete(hscss.store.highestChainBlockIndexKey)
+		if err != nil {
+			return err
+		}
+		hscss.store.cacheHighestChainBlockIndex = 0
+		return nil
+	}
+
+	newHighest := newLength - 1
+	err = dbTx.Put(hscss.store.highestChainBlockIndexKey, hscss.store.serializeIndex(newHighest))
 	if err != nil {
 		return err
 	}
 
-	hscss.store.cacheHighestChainBlockIndex = highestIndex
+	hscss.store.cacheHighestChainBlockIndex = newHighest
 
 	return nil
 }
@@ -83,5 +106,5 @@ func (hscss *headersSelectedChainStagingShard) isStaged() bool {
 	return len(hscss.addedByHash) != 0 ||
 		len(hscss.removedByHash) != 0 ||
 		len(hscss.addedByIndex) != 0 ||
-		len(hscss.addedByIndex) != 0
+		len(hscss.removedByIndex) != 0
 }
